+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed dimensions of the anchors in the rpn.py file.
+
+
 ## [0.2.6] - 2024-03-04
 
 ### Added